fix: conditional rendering of toggleButton and ui fixes for the same
